client: deprecate instructions_bincode
