+++ Cargo.toml
@@ -42,6 +42,5 @@ solana-pubkey = "2.2.1"
 solana-sdk = "2.2.1"
 solana-sdk-ids = "2.2.1"
 solana-short-vec = "2.2.1"
-solana-stake-interface = "1.2.1"
 solana-system-interface = "1.0.0"
 thiserror = "1.0.61"
