+++ clients/rust/Cargo.toml
@@ -10,7 +10,13 @@ edition = { workspace = true }
 
 [features]
 fetch = ["dep:solana-client", "dep:solana-sdk"]
-serde = ["dep:bincode", "dep:serde", "kaigan/serde"]
+serde = [
+    "dep:bincode", 
+    "dep:serde", 
+    "dep:solana-config-interface", 
+    "kaigan/serde", 
+    "solana-config-interface/bincode"
+]
 test-sbf = []
 
 [dependencies]
@@ -19,6 +25,7 @@ borsh = { workspace = true }
 kaigan = { workspace = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 solana-client = { workspace = true, optional = true }
+solana-config-interface = { workspace = true, optional = true }
 solana-program = { workspace = true, features = ["borsh"] }
 solana-sdk = { workspace = true, optional = true }
