+++ clients/rust/src/lib.rs
@@ -2,7 +2,7 @@
 
 mod generated;
 mod hooked;
-
+#[deprecated(since = "1.0.0", note = "use `solana_config_interface` crate instead")]
 #[cfg(feature = "serde")]
 pub mod instructions_bincode;
