+++ interface/src/instruction.rs
@@ -1,29 +1,29 @@
+//! Program instruction helpers.
+
 use {
-    crate::{
-        id,
-        state::{ConfigKeys, ConfigState},
-    },
+    crate::{id, state::ConfigKeys},
+    bincode::serialized_size,
     solana_instruction::{AccountMeta, Instruction},
     solana_pubkey::Pubkey,
-    solana_system_interface::instruction as system_instruction,
 };
 
-fn initialize_account<T: ConfigState>(config_pubkey: &Pubkey) -> Instruction {
+fn initialize_account<T: Default + serde::Serialize>(config_pubkey: &Pubkey) -> Instruction {
     let account_metas = vec![AccountMeta::new(*config_pubkey, true)];
     let account_data = (ConfigKeys { keys: vec![] }, T::default());
     Instruction::new_with_bincode(id(), &account_data, account_metas)
 }
 
 /// Create a new, empty configuration account
-pub fn create_account<T: ConfigState>(
+pub fn create_account_with_max_config_space<T: Default + serde::Serialize>(
     from_account_pubkey: &Pubkey,
     config_account_pubkey: &Pubkey,
     lamports: u64,
+    max_config_space: u64,
     keys: Vec<(Pubkey, bool)>,
 ) -> Vec<Instruction> {
-    let space = T::max_space() + bincode::serialized_size(&ConfigKeys { keys }).unwrap();
+    let space = max_config_space.saturating_add(serialized_size(&ConfigKeys { keys }).unwrap());
     vec![
-        system_instruction::create_account(
+        solana_system_interface::instruction::create_account(
             from_account_pubkey,
             config_account_pubkey,
             lamports,
@@ -35,7 +35,7 @@ pub fn create_account<T: ConfigState>(
 }
 
 /// Store new data in a configuration account
-pub fn store<T: ConfigState>(
+pub fn store<T: serde::Serialize>(
     config_account_pubkey: &Pubkey,
     is_config_signer: bool,
     keys: Vec<(Pubkey, bool)>,
