+++ interface/src/state.rs
@@ -1,30 +1,9 @@
 use solana_pubkey::Pubkey;
-#[cfg(feature = "bincode")]
-#[allow(deprecated)]
-use {
-    bincode::{deserialize, serialized_size},
-    solana_stake_interface::config::Config as StakeConfig,
-};
 #[cfg(feature = "serde")]
 use {
     serde_derive::{Deserialize, Serialize},
     solana_short_vec as short_vec,
 };
-
-#[cfg(feature = "serde")]
-pub trait ConfigState: serde::Serialize + Default {
-    /// Maximum space that the serialized representation will require
-    fn max_space() -> u64;
-}
-
-// TODO move ConfigState into `solana_program` to implement trait locally
-#[cfg(feature = "bincode")]
-#[allow(deprecated)]
-impl ConfigState for StakeConfig {
-    fn max_space() -> u64 {
-        serialized_size(&StakeConfig::default()).unwrap()
-    }
-}
 
 /// A collection of keys to be stored in Config account data.
 #[derive(Debug, Default)]
@@ -34,11 +13,4 @@ pub struct ConfigKeys {
     // and `bool` whether that key is a signer of the data
     #[cfg_attr(feature = "serde", serde(with = "short_vec"))]
     pub keys: Vec<(Pubkey, bool)>,
-}
-
-#[cfg(feature = "bincode")]
-pub fn get_config_data(bytes: &[u8]) -> Result<&[u8], bincode::Error> {
-    deserialize::<ConfigKeys>(bytes)
-        .and_then(|keys| serialized_size(&keys))
-        .map(|offset| &bytes[offset as usize..])
 }
