+++ program/Cargo.toml
@@ -27,7 +27,7 @@ thiserror = { workspace = true }
 [dev-dependencies]
 mollusk-svm = { workspace = true, features = ["fuzz-fd"] }
 mollusk-svm-bencher = { workspace = true }
-solana-config-program-client = { path = "../clients/rust", features = ["serde"] }
+solana-config-interface = { workspace = true, features = ["bincode", "serde"] }
 solana-sdk = { workspace = true }
 
 [lib]
